refactored dependencies to pyproject.toml
