+++ frogger/robots/robots.py
@@ -6,8 +6,8 @@ from pydrake.multibody.parsing import Parser
 from pydrake.multibody.plant import MultibodyPlant
 from pydrake.multibody.tree import Body, Frame
 
-from core.objects import ObjectDescription
-from core.robots.robot_core import RobotModel
+from frogger.objects import ObjectDescription
+from frogger.robots.robot_core import RobotModel
 
 current_dir = Path(__file__)
 project_dir = [p for p in current_dir.parents if p.parts[-1] == "manipulation"][0]
