+++ frogger/sdfs/smoothing.py
