+++ scripts/paper/fc_correlation.py
@@ -8,9 +8,9 @@ import numpy as np
 import trimesh_fork as trimesh
 from pydrake.math import RigidTransform, RotationMatrix
 
-from core.objects import MeshObject
-from core.paper_experiments import run_exp
-from core.robots.robots import FR3AlgrModel
+from frogger.objects import MeshObject
+from frogger.paper_experiments import run_exp
+from frogger.robots.robots import FR3AlgrModel
 
 """
 This script computes the correlation between the min-weight and ferrari-canny metrics.
