+++ scripts/paper/fig1_example.py
@@ -5,8 +5,8 @@ import numpy as np
 import trimesh_fork as trimesh
 from pydrake.math import RigidTransform, RotationMatrix
 
-from core.objects import MeshObject
-from core.robots.robots import FR3AlgrModel
+from frogger.objects import MeshObject
+from frogger.robots.robots import FR3AlgrModel
 
 """
 Specific grasp for figure 1. I manually screenshotted an angle for the fig.
