--- setup.py
@@ -1,23 +1,0 @@
-# flake8: noqa
-import setuptools
-
-with open("README.md", "r") as fh:
-    long_description = fh.read()
-
-setuptools.setup(
-    name="frogger",
-    version="0.0.1",
-    author="Albert H. Li",
-    author_email="alberthli@caltech.edu",
-    description="FRoGGeR: Fast Robust Grasp Generation via the Min-Weight Metric.",
-    long_description=long_description,
-    long_description_content_type="text/markdown",
-    url="https://github.com/alberthli/frogger",
-    packages=setuptools.find_packages(),
-    classifiers=[
-        "Programming Language :: Python :: 3",
-        "License :: OSI Approved :: MIT License",
-    ],
-    python_requires=">=3.10",
-    license="MIT",
-)
